+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.3
+ - Implement auth for the xbox 360
 ## 1.2
  - Several bug fixes and changes to make the portal emulation closer to the real portals
  - Use the flippers speaker to emulate the portals speaker